Allow to set float number as mirror timeout.
I noticed some inconsistency in mirror timeout setting during playing
with gemstash. Default value is `0.1`, but when I set timeout value to
`0.5` it behaved as with `0` value. It set proper timeout only when I
set `1` as timeout value. I think it is also not proper behaviour with
`0.1` default and while setting cast it to Fixnum.

+++ lib/bundler/mirror.rb
@@ -81,7 +81,7 @@ module Bundler
         when false, "false"
           @fallback_timeout = 0
         else
-          @fallback_timeout = timeout.to_i
+          @fallback_timeout = timeout.to_f
         end
         @valid = nil
       end
